refactor(redis): move RedisTestContainerConfig to @zeltjs/redis/testing subpath
Better colocation: Redis-related testing utilities now live in the redis package.
Import path changes from @zeltjs/testing/redis to @zeltjs/redis/testing.
testcontainers remains an optional peer dependency.

Verified: precommit-verify ✕
