+++ packages/redis/tsdown.config.ts
@@ -1,12 +1,19 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/testing/index.ts'],
   format: ['esm'],
   dts: true,
   clean: true,
   fixedExtension: false,
   deps: {
-    neverBundle: ['@zeltjs/core', /^@zeltjs\/core\//, 'ioredis'],
+    neverBundle: [
+      '@zeltjs/core',
+      /^@zeltjs\/core\//,
+      '@zeltjs/testing',
+      /^@zeltjs\/testing\//,
+      'ioredis',
+      'testcontainers',
+    ],
   },
 });
