+++ packages/testing/package.json
@@ -31,10 +31,6 @@
     "./node": {
       "types": "./dist/adapters/node.d.ts",
       "import": "./dist/adapters/node.js"
-    },
-    "./redis": {
-      "types": "./dist/redis/index.d.ts",
-      "import": "./dist/redis/index.js"
     }
   },
   "files": [
@@ -47,20 +43,12 @@
   },
   "peerDependencies": {
     "@zeltjs/core": "workspace:*",
-    "@zeltjs/redis": "workspace:*",
-    "testcontainers": ">=10 <12",
     "vitest": ">=2",
     "@jest/globals": ">=29",
     "bun-types": ">=1",
     "@types/node": ">=20"
   },
   "peerDependenciesMeta": {
-    "@zeltjs/redis": {
-      "optional": true
-    },
-    "testcontainers": {
-      "optional": true
-    },
     "vitest": {
       "optional": true
     },
@@ -75,10 +63,8 @@
     }
   },
   "devDependencies": {
-    "@zeltjs/redis": "workspace:*",
     "@jest/globals": "29.7.0",
     "bun-types": "1.2.14",
-    "testcontainers": "10.25.0",
     "vitest": "4.1.5"
   }
 }
